requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.7.0
pydantic[email]>=2.10.0
openai>=1.59.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
python-dateutil>=2.9.0
pytz>=2024.2
httpx>=0.28.0
websockets>=14.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
